tests(modkit-db): improve test coverage
